+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Add methods to start and stop a dispatcher's transports, inbounds, and
+  outbounds separately.
+- Add `NewFx{{Service}}YARPCClient` and `NewFx{{Service}}YARPCProcedures`
+  generated methods from protoc-gen-yarpc-go for Fx.
+
 ## [1.28.0] - 2018-04-13
 ### Changed
 - Enabled random shuffling of peerlist order by default.
@@ -913,6 +920,7 @@ This release requires regeneration of ThriftRW code.
 
 - Initial release.
 
+[Unreleased]: https://github.com/yarpc/yarpc-go/compare/v1.28.0...HEAD
 [1.28.0]: https://github.com/yarpc/yarpc-go/compare/v1.27.2...v1.28.0
 [1.27.2]: https://github.com/yarpc/yarpc-go/compare/v1.27.1...v1.27.2
 [1.27.1]: https://github.com/yarpc/yarpc-go/compare/v1.27.0...v1.27.1